requirements.txt
slack-bolt>=1.18.0
pyyaml>=6.0
python-dotenv>=1.0.0
notion-client>=2.2.1,<3.0.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langchain-google-genai>=2.0.0
